Move file operations into dedicated file
